tactyk-test restructure
