+++ tactyktest/ttest.c
@@ -19,15 +19,7 @@
 #include "tactyk_dblock.h"
 #include "tactyk_alloc.h"
 
-#define TACTYK_TESTSTATE__EXIT 0
-#define TACTYK_TESTSTATE__PASS 1
-#define TACTYK_TESTSTATE__FAIL 2
-#define TACTYK_TESTSTATE__TEST_ERROR 3
-#define TACTYK_TESTSTATE__RUNNING 4
-#define TACTYK_TESTSTATE__PREPARING 5
-#define TACTYK_TESTSTATE__INITIALIZING 6
-#define TACTYK_TESTSTATE__INACTIVE 7
-#define TACTYK_TESTSTATE__CONTINUE 8
+#include "ttest.h"
 
 #define TEST_CLASS_I8 1
 #define TEST_CLASS_I16 2
@@ -45,143 +37,6 @@
 #define OBJECT_VM 5
 #define OBJECT_VM_PROGRAM_DECLARATION 6
 
-struct tactyk_test_entry;
-union tactyk_test__value;
-
-typedef bool (tactyk_test__VALUE_ADJUSTER) (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-typedef uint64_t (tactyk_test__VALUE_TESTER) (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-
-// abstract test handler
-struct tactyk_test_entry {
-    char *name;
-    tactyk_test__VALUE_ADJUSTER *adjust;        // Setter to use during STATE opreation
-    tactyk_test__VALUE_TESTER *test;            // Test function to use for TEST operation
-    uint64_t element_offset;    // object member to access
-    uint64_t array_offset;      // If the object is located in an array, which object within the array to access
-};
-
-union tactyk_test__value {
-    uint8_t i8;
-    uint8_t i16;
-    uint8_t i32;
-    uint8_t i64;
-    float f32;
-    double f64;
-};
-
-typedef uint64_t (*tactyk_emit__test_func)(struct tactyk_dblock__DBlock *data);
-
-struct tactyk_test__Program {
-    struct tactyk_dblock__DBlock *name;
-    struct tactyk_pl__Context *plctx;
-    struct tactyk_asmvm__Program *program;
-};
-
-struct tactyk_asmvm__Program *tprg;
-struct tactyk_asmvm__Context *vmctx;
-
-// a binary copy of vmctx to be used to scan for unexpected state transitions.
-// This is to be updated by the tactyk_test any time tactyk_test checks or alters a property from the real vmctx.
-// Unexpected state transitions are to be captured by scanning vmctx and shadow_vmctx for any deviations.
-struct tactyk_asmvm__Context *shadow_vmctx;
-struct tactyk_asmvm__memblock_lowlevel *shadow_memblocks;
-struct tactyk_asmvm__MicrocontextStash *shadow_mctxstack;
-uint32_t *shadow_lwcall_stack;
-double precision;
-uint64_t callback_id;
-uint64_t ccall_args[6];
-int64_t ccall_retval;
-
-uint64_t tactyk_test__exec_test_commands(struct tactyk_dblock__DBlock *test_spec);
-uint64_t tactyk_test__next(struct tactyk_dblock__DBlock *test_spec);
-void tactyk_test__exit(uint64_t test_result);
-
-uint64_t tactyk_test__PROGRAM(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__EXEC(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__RECV_CCALL_1(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f);
-uint64_t tactyk_test__RECV_CCALL_2(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f);
-uint64_t tactyk_test__RECV_CCALL_3(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f);
-uint64_t tactyk_test__RECV_CCALL_4(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f);
-void tactyk_test__RECV_TCALL_5(struct tactyk_asmvm__Context *ctx);
-void tactyk_test__RECV_TCALL_6(struct tactyk_asmvm__Context *ctx);
-void tactyk_test__RECV_TCALL_7(struct tactyk_asmvm__Context *ctx);
-void tactyk_test__RECV_TCALL_8(struct tactyk_asmvm__Context *ctx);
-uint64_t tactyk_test__TEST(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__STATE(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__ALLOC(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__DATA(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__REF(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__ERROR(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__CONTINUE(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__RETURN(struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__RESUME(struct tactyk_dblock__DBlock *spec);
-
-bool tactyk_test__SET_CONTEXT_STATUS(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_CONTEXT_STATUS(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-
-void tactyk_test__mk_var_test(char *name, tactyk_test__VALUE_ADJUSTER setter, tactyk_test__VALUE_TESTER *tester);
-bool tactyk_test__SET_REGISTER (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-bool tactyk_test__SET_XMM_REGISTER_FLOAT (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-bool tactyk_test__SET_ADDR (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-bool tactyk_test__SET_MEM (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_REGISTER(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_XMM_REGISTER_FLOAT (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_ADDR(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_MEM(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_CALLBACK(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_CCALL_ARGUMENT(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_LWCALL_STACK(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-uint64_t tactyk_test__TEST_STASH(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec);
-void tactyk_test__mk_register_test(char *name, uint64_t ofs);
-void tactyk_test__mk_xmm_register_test(char *name, uint64_t ofs);
-void tactyk_test__mk_ccallarg_test(char *name, uint64_t ofs);
-
-void tactyk_test__report(char *msg);
-
-struct tactyk_dblock__DBlock *base_tests;
-
-struct tactyk_emit__Context *emitctx;
-struct tactyk_asmvm__VM *vm;
-struct tactyk_dblock__DBlock *programs;
-struct tactyk_dblock__DBlock *shadow_memblock_sets;
-struct tactyk_dblock__DBlock *test_functions;
-
-struct tactyk_dblock__DBlock *DEFAULT_NAME;
-
-double DEFAULT_PRECISION = DBL_EPSILON * 256.0;
-
-struct tactyk_dblock__DBlock **active_test_spec;
-
-// test command immediately following a "CONTINUE" command
-// This is used by callbacks [from tactyk] to update test-command control flow so that multiple callbacks can be handled.
-struct tactyk_dblock__DBlock *test_continuation;
-
-#define TACTYK_TEST__FNAME_BUFSIZE (1<<10)
-#define TACTYK_TEST__REPORT_BUFSIZE (1<<10)
-#define TACTYK_TEST__DUMP_BUFSIZE (1<<14)
-struct tactyk_test__Status {
-    uint64_t pid;
-    uint64_t testid;
-    time_t start_time;
-    double max_age;
-    uint64_t test_result;
-    bool ran;
-    char dump_context_expectation[TACTYK_TEST__DUMP_BUFSIZE];
-    char dump_stack_expectation[TACTYK_TEST__DUMP_BUFSIZE];
-    char dump_context_observed[TACTYK_TEST__DUMP_BUFSIZE];
-    char dump_stack_observed[TACTYK_TEST__DUMP_BUFSIZE];
-    char fname[TACTYK_TEST__FNAME_BUFSIZE];
-    char error[TACTYK_TEST__REPORT_BUFSIZE];
-    char warning[TACTYK_TEST__REPORT_BUFSIZE];
-    char report[TACTYK_TEST__REPORT_BUFSIZE];
-    //uint64_t test_src_size;
-    //char test_data[TACTYK_TEST__SRC_BUFSIZE];
-};
-
-void tactyk_test__prepare(struct tactyk_test__Status *tstate);
-void tactyk_test__run(struct tactyk_test__Status *tstate);
-void tactyk_test__reset_state(struct tactyk_test__Status *tstate);
-void tactyk_test__await_start(struct tactyk_test__Status *tstate);
 
 uint64_t test_count;
 uint64_t tests_started;
@@ -390,7 +245,7 @@ int main(int argc, char *argv[], char *envp[]) {
                 case TACTYK_TESTSTATE__INITIALIZING:
                 case TACTYK_TESTSTATE__RUNNING: {
                     int status = 0;
-                    pid_t pid = waitpid(tstate->pid, &status, WNOHANG);
+                    waitpid(tstate->pid, &status, WNOHANG);
                     if (WIFSIGNALED(status)) {
                         int sig = WTERMSIG(status);
                         failed += 1;
@@ -507,2032 +362,6 @@ void tactyk_test__await_start(struct tactyk_test__Status *tstate) {
     }
 }
 
-//static jmp_buf tactyk_test_err_jbuf;
 
-struct tactyk_dblock__DBlock *ERROR_TOKEN;
-struct tactyk_dblock__DBlock *WARNING_TOKEN;
-struct tactyk_test__Status *test_state;
 
-void tactyk_test__warning_handler(char *msg, void *data) {
-    memset(test_state->warning, 0, TACTYK_TEST__REPORT_BUFSIZE);
-    strncpy(test_state->warning, msg, TACTYK_TEST__REPORT_BUFSIZE);
-    int64_t space_remaining = TACTYK_TEST__REPORT_BUFSIZE - strlen(msg)-3;
 
-    if ((space_remaining > 0) && (data != NULL)) {
-        test_state->warning[strlen(msg)+0] = ':';
-        test_state->warning[strlen(msg)+1] = ' ';
-        tactyk_dblock__export_cstring(&test_state->warning[strlen(msg)+2], space_remaining, data);
-    }
-}
-
-void tactyk_test__error_handler(char *msg, void *data) {
-    memset(test_state->error, 0, TACTYK_TEST__REPORT_BUFSIZE);
-    strncpy(test_state->error, msg, TACTYK_TEST__REPORT_BUFSIZE);
-    int64_t space_remaining = TACTYK_TEST__REPORT_BUFSIZE - strlen(msg)-3;
-
-    if ((space_remaining > 0) && (data != NULL)) {
-        test_state->error[strlen(msg)+0] = ':';
-        test_state->error[strlen(msg)+1] = ' ';
-        tactyk_dblock__export_cstring(&test_state->error[strlen(msg)+2], space_remaining, data);
-    }
-
-    //longjmp(tactyk_test_err_jbuf, 1);
-    struct tactyk_dblock__DBlock *test_spec = *active_test_spec;
-    if (test_spec == NULL) {
-        assert(test_state->test_result == TACTYK_TESTSTATE__INITIALIZING);
-        sprintf(test_state->report, "Unexpected error during test initialization.");
-        tactyk_test__exit(TACTYK_TESTSTATE__TEST_ERROR);
-    }
-    else {
-        test_spec = test_spec->next;
-    }
-
-    if (test_spec != NULL) {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, test_spec->token);
-        if (strncmp(buf, "ERROR", 6) == 0) {
-            tactyk_test__exit(TACTYK_TESTSTATE__PASS);
-        }
-        // unexpected errors are a test failure.
-        tactyk_test__report("Unexpected error");
-        tactyk_test__exit(TACTYK_TESTSTATE__FAIL);
-    }
-    else {
-        // failure at the end of the test
-        //  (only expected to occur if a test ends with an EXEC command it triggers an error)
-        tactyk_test__report("Unexpected error at end of test");
-        tactyk_test__exit(TACTYK_TESTSTATE__FAIL);
-    }
-}
-
-void tactyk_test__run(struct tactyk_test__Status *tstate) {
-    test_state = tstate;
-    tstate->test_result = TACTYK_TESTSTATE__INITIALIZING;
-
-    FILE *f = fopen(tstate->fname, "r");
-    fseek(f, 0, SEEK_END);
-    int64_t len = ftell(f);
-    char *src = calloc(len+1, sizeof(char));
-    fseek(f,0, SEEK_SET);
-    fread(src, len, 1, f);
-    fclose(f);
-
-    active_test_spec = calloc(1, sizeof(void*));
-    *active_test_spec = NULL;
-    test_continuation = NULL;
-
-    tactyk_init();
-
-    //tstate->test_result = TACTYK_TESTSTATE__PASS;
-    //_exit(0);
-    error = tactyk_test__error_handler;
-    warn = tactyk_test__warning_handler;
-
-    ERROR_TOKEN = tactyk_dblock__from_c_string("ERROR");
-    WARNING_TOKEN = tactyk_dblock__from_c_string("WARNING");
-
-    //if (setjmp(tactyk_err_jbuf)) {
-    //    _exit(0);
-    //}
-
-    // should make configuration mutable - some tests will need to load alternative virtual ISA specifications that include special instructions which test
-    //      the integrity of the sandbox.
-    tactyk_visa__init("rsc/tactyk_core.visa");
-    emitctx = tactyk_emit__init();
-    emitctx->visa_file_prefix = "rsc/";
-
-    tactyk_visa__init_emit(emitctx);
-    tactyk_pl__init();
-
-    vm = tactyk_asmvm__new_vm();
-    vmctx = tactyk_asmvm__new_context(vm);
-    shadow_vmctx = calloc(1, sizeof(struct tactyk_asmvm__Context));
-    shadow_mctxstack = (struct tactyk_asmvm__MicrocontextStash*) calloc(TACTYK_ASMVM__MCTX_STACK_SIZE, sizeof(struct tactyk_asmvm__MicrocontextStash));
-    shadow_lwcall_stack = (uint32_t*) calloc(TACTYK_ASMVM__LWCALL_STACK_SIZE, sizeof(uint32_t));
-    tactyk_debug__configure_api(emitctx);
-    tactyk_emit_svc__configure(emitctx);
-    tactyk_emit__add_c_apifunc(emitctx, "cfunc-1", tactyk_test__RECV_CCALL_1);
-    tactyk_emit__add_c_apifunc(emitctx, "cfunc-2", tactyk_test__RECV_CCALL_2);
-    tactyk_emit__add_c_apifunc(emitctx, "cfunc-3", tactyk_test__RECV_CCALL_3);
-    tactyk_emit__add_c_apifunc(emitctx, "cfunc-4", tactyk_test__RECV_CCALL_4);
-    tactyk_emit__add_tactyk_apifunc(emitctx, "tfunc-5", tactyk_test__RECV_TCALL_5);
-    tactyk_emit__add_tactyk_apifunc(emitctx, "tfunc-6", tactyk_test__RECV_TCALL_6);
-    tactyk_emit__add_tactyk_apifunc(emitctx, "tfunc-7", tactyk_test__RECV_TCALL_7);
-    tactyk_emit__add_tactyk_apifunc(emitctx, "tfunc-8", tactyk_test__RECV_TCALL_8);
-
-    programs = tactyk_dblock__new_table(64);
-    shadow_memblock_sets = tactyk_dblock__new_table(64);
-    DEFAULT_NAME = tactyk_dblock__from_safe_c_string("DEFAULT");
-    precision = DEFAULT_PRECISION;
-    test_functions = tactyk_dblock__new_table(64);
-
-    tactyk_dblock__put(test_functions, "PROGRAM", tactyk_test__PROGRAM);
-    tactyk_dblock__put(test_functions, "EXEC", tactyk_test__EXEC);
-    tactyk_dblock__put(test_functions, "TEST", tactyk_test__TEST);
-    tactyk_dblock__put(test_functions, "STATE", tactyk_test__STATE);
-    tactyk_dblock__put(test_functions, "ALLOC", tactyk_test__ALLOC);
-    tactyk_dblock__put(test_functions, "DATA", tactyk_test__DATA);
-    tactyk_dblock__put(test_functions, "REF", tactyk_test__REF);
-    tactyk_dblock__put(test_functions, "CONTINUE", tactyk_test__CONTINUE);
-    tactyk_dblock__put(test_functions, "RETURN", tactyk_test__RETURN);
-    tactyk_dblock__put(test_functions, "RESUME", tactyk_test__RESUME);
-
-    base_tests = tactyk_dblock__new_managedobject_table(1024, sizeof(struct tactyk_test_entry));
-    tactyk_test__mk_var_test("status", tactyk_test__SET_CONTEXT_STATUS, tactyk_test__TEST_CONTEXT_STATUS);
-    tactyk_test__mk_var_test("status", tactyk_test__SET_CONTEXT_STATUS, tactyk_test__TEST_CONTEXT_STATUS);
-    tactyk_test__mk_var_test("status", tactyk_test__SET_CONTEXT_STATUS, tactyk_test__TEST_CONTEXT_STATUS);
-    tactyk_test__mk_var_test("status", tactyk_test__SET_CONTEXT_STATUS, tactyk_test__TEST_CONTEXT_STATUS);
-    tactyk_test__mk_var_test("status", tactyk_test__SET_CONTEXT_STATUS, tactyk_test__TEST_CONTEXT_STATUS);
-
-    struct tactyk_test_entry *addr1_test = tactyk_dblock__new_managedobject(base_tests, "addr1");
-    addr1_test->name = "addr1";
-    addr1_test->adjust = tactyk_test__SET_ADDR;
-    addr1_test->test = tactyk_test__TEST_ADDR;
-    addr1_test->element_offset = 1;
-    addr1_test->array_offset = 0;
-    struct tactyk_test_entry *addr2_test = tactyk_dblock__new_managedobject(base_tests, "addr2");
-    addr2_test->name = "addr2";
-    addr2_test->adjust = tactyk_test__SET_ADDR;
-    addr2_test->test = tactyk_test__TEST_ADDR;
-    addr2_test->element_offset = 2;
-    addr2_test->array_offset = 0;
-    struct tactyk_test_entry *addr3_test = tactyk_dblock__new_managedobject(base_tests, "addr3");
-    addr3_test->name = "addr3";
-    addr3_test->adjust = tactyk_test__SET_ADDR;
-    addr3_test->test = tactyk_test__TEST_ADDR;
-    addr3_test->element_offset = 3;
-    addr3_test->array_offset = 0;
-    struct tactyk_test_entry *addr4_test = tactyk_dblock__new_managedobject(base_tests, "addr4");
-    addr4_test->name = "addr4";
-    addr4_test->adjust = tactyk_test__SET_ADDR;
-    addr4_test->test = tactyk_test__TEST_ADDR;
-    addr4_test->element_offset = 4;
-    addr4_test->array_offset = 0;
-
-    struct tactyk_test_entry *mem_test = tactyk_dblock__new_managedobject(base_tests, "mem");
-    mem_test->name = "mem";
-    mem_test->adjust = tactyk_test__SET_MEM;
-    mem_test->test = tactyk_test__TEST_MEM;
-    mem_test->element_offset = 0;
-    mem_test->array_offset = 0;
-
-    struct tactyk_test_entry *cb_test = tactyk_dblock__new_managedobject(base_tests, "callback");
-    cb_test->name = "callback";
-    cb_test->adjust = NULL;
-    cb_test->test = tactyk_test__TEST_CALLBACK;
-    cb_test->element_offset = 0;
-    cb_test->array_offset = 0;
-
-    struct tactyk_test_entry *stash_test = tactyk_dblock__new_managedobject(base_tests, "stash");
-    stash_test->name = "stash";
-    stash_test->adjust = NULL;
-    stash_test->test = tactyk_test__TEST_STASH;
-    stash_test->element_offset = 0;
-    stash_test->array_offset = 0;
-
-    struct tactyk_test_entry *lwcs_test = tactyk_dblock__new_managedobject(base_tests, "lwcstack");
-    lwcs_test->name = "lwcstack";
-    lwcs_test->adjust = NULL;
-    lwcs_test->test = tactyk_test__TEST_LWCALL_STACK;
-    lwcs_test->element_offset = 0;
-    lwcs_test->array_offset = 0;
-
-    tactyk_test__mk_register_test("rLWCSI", 1);
-    tactyk_test__mk_register_test("rMCSI", 2);
-    tactyk_test__mk_register_test("rA", 10);
-    tactyk_test__mk_register_test("rB", 11);
-    tactyk_test__mk_register_test("rC", 12);
-    tactyk_test__mk_register_test("rD", 13);
-    tactyk_test__mk_register_test("rE", 14);
-    tactyk_test__mk_register_test("rF", 15);
-    tactyk_test__mk_xmm_register_test("xA", 0);
-    tactyk_test__mk_xmm_register_test("xB", 1);
-    tactyk_test__mk_xmm_register_test("xC", 2);
-    tactyk_test__mk_xmm_register_test("xD", 3);
-    tactyk_test__mk_xmm_register_test("xE", 4);
-    tactyk_test__mk_xmm_register_test("xF", 5);
-    tactyk_test__mk_xmm_register_test("xG", 6);
-    tactyk_test__mk_xmm_register_test("xH", 7);
-    tactyk_test__mk_xmm_register_test("xI", 8);
-    tactyk_test__mk_xmm_register_test("xJ", 9);
-    tactyk_test__mk_xmm_register_test("xK", 10);
-    tactyk_test__mk_xmm_register_test("xL", 11);
-    tactyk_test__mk_xmm_register_test("xM", 12);
-    tactyk_test__mk_xmm_register_test("xN", 13);
-    tactyk_test__mk_xmm_register_test("xTEMPA", 14);
-    tactyk_test__mk_xmm_register_test("xTEMPB", 15);
-
-    tactyk_test__mk_ccallarg_test("arg0", 0);
-    tactyk_test__mk_ccallarg_test("arg1", 1);
-    tactyk_test__mk_ccallarg_test("arg2", 2);
-    tactyk_test__mk_ccallarg_test("arg3", 3);
-    tactyk_test__mk_ccallarg_test("arg4", 4);
-    tactyk_test__mk_ccallarg_test("arg5", 5);
-
-    //struct tactyk_dblock__DBlock *test_src = tactyk_dblock__from_bytes(NULL, fbytes, 0, (uint64_t)len, true);
-    struct tactyk_dblock__DBlock *test_src = tactyk_dblock__from_safe_c_string(src);
-    tactyk_dblock__fix(test_src);
-    tactyk_dblock__tokenize(test_src, '\n', false);
-    struct tactyk_dblock__DBlock *test_spec = tactyk_dblock__remove_blanks(test_src, ' ', '#');
-    tactyk_dblock__stratify(test_spec, ' ');
-    tactyk_dblock__trim(test_spec);
-    tactyk_dblock__tokenize(test_spec, ' ', true);
-    tactyk_dblock__set_persistence_code(test_spec, 100);
-
-    tactyk_dblock__set_persistence_code(test_src, 1000);
-    tactyk_dblock__set_persistence_code(test_spec, 1000);
-    tactyk_dblock__set_persistence_code(base_tests, 1000);
-    tactyk_dblock__set_persistence_code(test_functions, 1000);
-    tactyk_dblock__set_persistence_code(programs, 1000);
-    tactyk_dblock__set_persistence_code(shadow_memblock_sets, 1000);
-
-    *active_test_spec = test_spec;
-    tstate->test_result = TACTYK_TESTSTATE__RUNNING;
-
-    ccall_retval = 0;
-    callback_id = 0;
-    memset(ccall_args, 0, sizeof(ccall_args));
-
-    uint64_t test_result = tactyk_test__exec_test_commands(test_spec);
-    tactyk_test__exit(test_result);
-}
-
-uint64_t tactyk_test__exec_test_commands(struct tactyk_dblock__DBlock *test_spec) {
-    while (test_spec != NULL) {
-        if (test_continuation != NULL) {
-            tactyk_test__report("Expected callback not received");
-            tactyk_test__exit(TACTYK_TESTSTATE__FAIL);
-        }
-        uint64_t tresult = tactyk_test__next(test_spec);
-        switch(tresult) {
-            case TACTYK_TESTSTATE__PASS: {
-                break;
-            }
-            case TACTYK_TESTSTATE__FAIL: {
-                tactyk_test__exit(TACTYK_TESTSTATE__FAIL);
-                break;
-            }
-            case TACTYK_TESTSTATE__EXIT: {
-                tactyk_test__exit(TACTYK_TESTSTATE__PASS);
-                break;
-            }
-            case TACTYK_TESTSTATE__CONTINUE: {
-                test_continuation = test_spec->next;
-                return TACTYK_TESTSTATE__CONTINUE;
-            }
-            // impossible condition
-            default: {
-                tactyk_test__exit(TACTYK_TESTSTATE__TEST_ERROR);
-                break;
-            }
-        }
-        test_spec = test_spec->next;
-    }
-    return TACTYK_TESTSTATE__PASS;
-}
-
-void tactyk_test__report(char *msg) {
-    memset(test_state->report, 0, TACTYK_TEST__REPORT_BUFSIZE);
-    snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE-1, "%s", msg);
-}
-
-void tactyk_test__exit(uint64_t test_result) {
-    if (test_state->ran == true) {
-        FILE *stream = fmemopen(test_state->dump_context_observed, TACTYK_TEST__DUMP_BUFSIZE, "w");
-        tactyk_debug__write_context(vmctx, stream);
-        fflush(stream);
-        fclose(stream);
-
-        stream = fmemopen(test_state->dump_context_expectation, TACTYK_TEST__DUMP_BUFSIZE, "w");
-        tactyk_debug__write_context(shadow_vmctx, stream);
-        fflush(stream);
-        fclose(stream);
-
-        stream = fmemopen(test_state->dump_stack_observed, TACTYK_TEST__DUMP_BUFSIZE, "w");
-        tactyk_debug__write_vmstack(vmctx, stream);
-        fflush(stream);
-        fclose(stream);
-
-        stream = fmemopen(test_state->dump_stack_expectation, TACTYK_TEST__DUMP_BUFSIZE, "w");
-        tactyk_debug__write_vmstack(shadow_vmctx, stream);
-        fflush(stream);
-        fclose(stream);
-    }
-
-    test_state->test_result = test_result;
-    _exit(0);
-}
-
-// perform the next action.
-// This is called both by the while loop in the main function, and by test callbacks.
-//      (the use by test callbacks is for handling RECV tests without either complicating the test framework or disrupting control flow)
-uint64_t tactyk_test__next(struct tactyk_dblock__DBlock *test_spec) {
-    uint64_t tresult = 0;
-    struct tactyk_dblock__DBlock *token = test_spec->token;
-    if (token != NULL) {
-        tactyk_emit__test_func tfunc = tactyk_dblock__get(test_functions, test_spec->token);
-        if (tfunc == NULL) {
-            char buf[64];
-            tactyk_dblock__export_cstring(buf, 64, token);
-            sprintf(test_state->report, "Undefined test function: %s\n", buf);
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-        else {
-            *active_test_spec = test_spec;
-            tresult = tfunc(test_spec);
-            *active_test_spec = test_spec;
-            return tresult;
-        }
-    }
-    else {
-        return TACTYK_TESTSTATE__EXIT;
-    }
-}
-
-uint64_t tactyk_test__PROGRAM(struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *name = spec->token->next;
-    if (name == NULL) {
-        name = DEFAULT_NAME;
-    }
-    struct tactyk_pl__Context *plctx = tactyk_pl__new(emitctx);
-    tactyk_pl__load_dblock(plctx, spec->child);
-    tprg = tactyk_pl__build(plctx);
-    tactyk_asmvm__add_program(vmctx, tprg);
-
-    vmctx->memblocks = (struct tactyk_asmvm__memblock_lowlevel*) tprg->memory_layout_ll->data;
-    shadow_memblocks = (struct tactyk_asmvm__memblock_lowlevel*) calloc(TACTYK_ASMVM__MEMBLOCK_CAPACITY, sizeof(struct tactyk_asmvm__memblock_lowlevel));
-    memcpy(shadow_memblocks, vmctx->memblocks, TACTYK_ASMVM__MEMBLOCK_CAPACITY * sizeof(struct tactyk_asmvm__memblock_lowlevel));
-    for (uint64_t i = 0; i < TACTYK_ASMVM__MEMBLOCK_CAPACITY; i += 1) {
-        struct tactyk_asmvm__memblock_lowlevel *mll = &vmctx->memblocks[i];
-        struct tactyk_asmvm__memblock_lowlevel *shadow_mll = &shadow_memblocks[i];
-
-        if (mll->base_address != NULL) {
-            uint64_t sz = mll->array_bound + mll->element_bound -1;
-            uint8_t *shadow_bytes = calloc(1, sz);
-            shadow_mll->base_address = memcpy(shadow_bytes, mll->base_address, sz);
-        }
-    }
-    tactyk_dblock__put(shadow_memblock_sets, name, shadow_memblocks);
-
-    tactyk_dblock__put(programs, name, tprg);
-
-    return TACTYK_TESTSTATE__PASS;
-}
-
-uint64_t tactyk_test__EXEC(struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *program_name;
-    struct tactyk_dblock__DBlock *func_name = spec->token->next;
-    test_continuation = spec->child;
-    if (tactyk_dblock__count_tokens(spec) == 3) {
-        program_name = func_name;
-        func_name = func_name->next;
-        tprg = tactyk_dblock__get(programs, program_name);
-        shadow_memblocks = tactyk_dblock__get(shadow_memblock_sets, program_name);
-        if (tprg == NULL) {
-            char buf[64];
-            tactyk_dblock__export_cstring(buf, 64, program_name);
-            snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "EXEC -- program '%s' is undefined", buf);
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-    }
-    if ( (vmctx == NULL) || (tprg == NULL) ) {
-        tactyk_test__report("Program not built");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    if (func_name == NULL) {
-        tactyk_asmvm__prepare_invoke(shadow_vmctx, tprg, "MAIN");
-        tactyk_asmvm__call(vmctx, tprg, "MAIN");
-    }
-    else {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, func_name);
-
-        memcpy(shadow_vmctx, vmctx, sizeof(struct tactyk_asmvm__Context));
-        if (!tactyk_asmvm__prepare_invoke(shadow_vmctx, tprg, buf)) {
-            char bufpn[64];
-            tactyk_dblock__export_cstring(bufpn, 64, program_name);
-            char buffn[64];
-            tactyk_dblock__export_cstring(buffn, 64, func_name);
-            snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "EXEC -- Can not call invalid function '%s.%s'", bufpn, buffn);
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-        tactyk_asmvm__call(vmctx, tprg, buf);
-    }
-    // By default, expect the program to exit normally (by placing the 'STATUS_HALT' code in the shadow context)
-
-    shadow_vmctx->STATUS = 4;
-    test_state->ran = true;
-    return TACTYK_TESTSTATE__PASS;
-}
-
-
-uint64_t tactyk_test__RESUME(struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *program_name = spec->token->next;
-    if (program_name != NULL) {
-        tprg = tactyk_dblock__get(programs, program_name);
-        if (tprg == NULL) {
-            char bufpn[64];
-            tactyk_dblock__export_cstring(bufpn, 64, program_name);
-            snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "EXEC -- Can not resume undefined program '%s'", bufpn);
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-    }
-
-    tactyk_asmvm__resume(vmctx);
-    shadow_vmctx->STATUS = 4;
-    return TACTYK_TESTSTATE__PASS;
-}
-uint64_t tactyk_test__RECV_CCALL_1(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f) {
-    ccall_args[0] = a;
-    ccall_args[1] = b;
-    ccall_args[2] = c;
-    ccall_args[3] = d;
-    ccall_args[4] = e;
-    ccall_args[5] = f;
-    callback_id = 1;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-    return ccall_retval;
-}
-uint64_t tactyk_test__RECV_CCALL_2(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f) {
-    ccall_args[0] = a;
-    ccall_args[1] = b;
-    ccall_args[2] = c;
-    ccall_args[3] = d;
-    ccall_args[4] = e;
-    ccall_args[5] = f;
-    callback_id = 2;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-    return ccall_retval;
-}
-uint64_t tactyk_test__RECV_CCALL_3(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f) {
-    ccall_args[0] = a;
-    ccall_args[1] = b;
-    ccall_args[2] = c;
-    ccall_args[3] = d;
-    ccall_args[4] = e;
-    ccall_args[5] = f;
-    callback_id = 3;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-    return ccall_retval;
-}
-uint64_t tactyk_test__RECV_CCALL_4(int64_t a, int64_t b, int64_t c, int64_t d, int64_t e, int64_t f) {
-    ccall_args[0] = a;
-    ccall_args[1] = b;
-    ccall_args[2] = c;
-    ccall_args[3] = d;
-    ccall_args[4] = e;
-    ccall_args[5] = f;
-    callback_id = 4;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-    return ccall_retval;
-}
-void tactyk_test__RECV_TCALL_5(struct tactyk_asmvm__Context *ctx) {
-    callback_id = 5;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-}
-void tactyk_test__RECV_TCALL_6(struct tactyk_asmvm__Context *ctx) {
-    callback_id = 6;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-}
-void tactyk_test__RECV_TCALL_7(struct tactyk_asmvm__Context *ctx) {
-    callback_id = 7;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-}
-void tactyk_test__RECV_TCALL_8(struct tactyk_asmvm__Context *ctx) {
-    callback_id = 8;
-    test_continuation = NULL;
-    struct tactyk_dblock__DBlock *spec = *active_test_spec;
-    assert(spec != NULL);
-    if (spec->child != NULL) {
-        tactyk_test__exec_test_commands(spec->child);
-        spec->child = test_continuation;
-    }
-    *active_test_spec = spec;
-}
-
-bool tactyk_test__approximately_eq(double a, double b) {
-    if ( a == b ) {
-        return true;
-    }
-    else if (!isfinite(a)) {
-        return false;
-    }
-    else if (!isfinite(b)) {
-        return false;
-    }
-    else if (isnan(a)) {
-        return isnan(b);    // For state transition tracking purposes (NAN == NAN) is TRUE
-    }
-    else if (isnan(b)) {
-        return false;
-    }
-    else if (a == 0) {
-        return fabs(b) < precision;
-    }
-    else if (b == 0) {
-        return fabs(a) < precision;
-    }
-    return ( fabs(1.0-(a/b)) < precision);
-}
-
-uint64_t tactyk_test__TEST(struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *td = spec->child;
-    while (td != NULL) {
-        struct tactyk_dblock__DBlock *item_name = td->token;
-        assert(item_name != NULL);
-        struct tactyk_dblock__DBlock *item_value = td->token->next;
-
-        if (item_value == NULL) {
-            tactyk_test__report("Unspecified item value");
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-
-        struct tactyk_test_entry *test = tactyk_dblock__get(base_tests, item_name);
-        if (test == NULL) {
-            char buf[256];
-            sprintf(buf, "No handler for test item: ");
-            tactyk_dblock__export_cstring(&buf[strlen(buf)], 256-strlen(buf), item_name );
-            tactyk_test__report(buf);
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-
-        uint64_t tresult = test->test(test, td);
-        if (tresult != TACTYK_TESTSTATE__PASS) {
-            return tresult;
-        }
-
-        td = td->next;
-    }
-
-    // if callback id not cleared, then it means a ccall or tcall is not accounted for
-    //   this may occur either in a TEST command nested inside of an EXEC and invoked through a callback, or it may occur after the return form the callback when the next TEST command is handled.
-    if (callback_id != 0) {
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "Unexpected callback, id:%ju", callback_id);
-        return TACTYK_TESTSTATE__FAIL;
-    }
-
-    // not part of the test.
-    shadow_vmctx->instruction_index = vmctx->instruction_index;
-
-    uint8_t *real_bytes = (uint8_t*) vmctx;
-    uint8_t *shadow_bytes = (uint8_t*) shadow_vmctx;
-    uint64_t *real_qwords = (uint64_t*) vmctx;
-    uint64_t *shadow_qwords = (uint64_t*) shadow_vmctx;
-
-    // scan the context structure, but leave out runtime registers and diagnostic data.
-    //      diagnostic data is not part of testing (It's intended to aid with debugging)
-    //      Runtime registers is whatever C leaves on the registers when calling into TACTYK.
-    //          This would better be handled through a runtime environment correctness test to be performed within callbacks and/or after returning from tactyk.
-    for (int64_t ofs = 0; ofs < offsetof(struct tactyk_asmvm__Context, runtime_registers); ofs += 1) {
-
-        // if a deviation is found, the test fails.
-        if (real_bytes[ofs] != shadow_bytes[ofs]) {
-            int64_t qwpos = ofs / 8;
-
-            // Check offset-ranges to identify which section of the context data structure the deviation was found in.
-
-            // general context data.
-            if (ofs < offsetof(struct tactyk_asmvm__Context, reg)) {
-                snprintf(
-                    test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                    "Context-state deviation at offset %jd (qword #%jd), expected=%jd observed=%jd",
-                    ofs, qwpos, shadow_qwords[qwpos], real_qwords[qwpos]
-                );
-            }
-            else {
-                int64_t rel_ofs = ofs - offsetof(struct tactyk_asmvm__Context, reg);
-                // main register file (standard x86 registers)
-                if (rel_ofs < offsetof(struct tactyk_asmvm__register_bank, xa)) {
-                    snprintf(
-                        test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                        "Register-file deviation at offset %jd (qword #%jd), expected=%jd observed=%jd",
-                        rel_ofs, rel_ofs/8, shadow_qwords[qwpos], real_qwords[qwpos]
-                    );
-                }
-
-                // xmm register file
-                else {
-                    int64_t rel_ofs = ofs - offsetof(struct tactyk_asmvm__Context, reg) - offsetof(struct tactyk_asmvm__register_bank, xa);
-                    snprintf(
-                        test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                        "XMM Register-file deviation at offset %jd (qword #%jd), expected=%f observed=%f",
-                        rel_ofs, rel_ofs/8, *((double*)&shadow_qwords[qwpos]), *((double*)&real_qwords[qwpos])
-                    );
-                }
-            }
-            return TACTYK_TESTSTATE__FAIL;
-        }
-    }
-
-    for (uint64_t i = 0; i < TACTYK_ASMVM__MEMBLOCK_CAPACITY; i++) {
-        struct tactyk_asmvm__memblock_lowlevel *mbll = &vmctx->memblocks[i];
-        struct tactyk_asmvm__memblock_lowlevel *shadow_mbll = &shadow_memblocks[i];
-        //printf("p1=%p p2=%p\n", mbll, shadow_mbll);
-        if (mbll->array_bound != shadow_mbll->array_bound) {
-            snprintf(
-                test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                "memblock %ju: array_bound deviation, expected=%u observed=%u",
-                i, shadow_mbll->array_bound, mbll->array_bound
-            );
-        }
-        if (mbll->element_bound != shadow_mbll->element_bound) {
-            snprintf(
-                test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                "memblock %ju: element_bound deviation, expected=%u observed=%u",
-                i, shadow_mbll->element_bound, mbll->element_bound
-            );
-        }
-        if (mbll->memblock_index != shadow_mbll->memblock_index) {
-            snprintf(
-                test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                "memblock %ju: memblock_index deviation, expected=%u observed=%u",
-                i, shadow_mbll->memblock_index, mbll->memblock_index
-            );
-        }
-        if (mbll->offset != shadow_mbll->offset) {
-            snprintf(
-                test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                "memblock %ju: offset-field deviation, expected=%u observed=%u",
-                i, shadow_mbll->offset, mbll->offset
-            );
-        }
-        uint64_t len = mbll->array_bound + mbll->element_bound -1;
-        if (mbll->base_address != NULL) {
-            for (uint64_t j = 0; j < len; j += 1) {
-                if (mbll->base_address[j] != shadow_mbll->base_address[j]) {
-                    snprintf(
-                        test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                        "Memblock %ju data deviation at offset %ju, expected=%u observed=%u",
-                        i, j, shadow_mbll->base_address[j], mbll->base_address[j]
-                    );
-                    return TACTYK_TESTSTATE__FAIL;
-                }
-            }
-        }
-    }
-
-    for (uint64_t i = 0; i < TACTYK_ASMVM__LWCALL_STACK_SIZE; i++) {
-        if (vmctx->lwcall_stack[i] != shadow_lwcall_stack[i]) {
-            snprintf(
-                test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                "lwcall stack deviation at offset %ju: expected=%u observed=%u",
-                i, shadow_lwcall_stack[i], vmctx->lwcall_stack[i]
-            );
-            return TACTYK_TESTSTATE__FAIL;
-        }
-    }
-
-    for (uint64_t i = 0; i < TACTYK_ASMVM__MCTX_STACK_SIZE; i++) {
-        struct tactyk_asmvm__MicrocontextStash *stash = &vmctx->microcontext_stack[i];
-        struct tactyk_asmvm__MicrocontextStash *shadow_stash = &shadow_mctxstack[i];
-
-        for (uint64_t j = 0; j < 4; j++) {
-            struct tactyk_asmvm__memblock_lowlevel *mbll = &stash->memblocks[i];
-            struct tactyk_asmvm__memblock_lowlevel *shadow_mbll = &stash->memblocks[i];
-            if (mbll->base_address != shadow_mbll->base_address) {
-                snprintf(
-                    test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                    "stash[%ju] deviation - memblock[%ju].base_address: expected=%p, observed=%p",
-                    i, j, shadow_mbll->base_address, mbll->base_address
-                );
-                return TACTYK_TESTSTATE__FAIL;
-            }
-            if (mbll->array_bound != shadow_mbll->array_bound) {
-                snprintf(
-                    test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                    "stash[%ju] deviation - memblock[%ju].array_bound: expected=%u, observed=%u",
-                    i, j, shadow_mbll->array_bound, mbll->array_bound
-                );
-                return TACTYK_TESTSTATE__FAIL;
-            }
-            if (mbll->element_bound != shadow_mbll->element_bound) {
-                snprintf(
-                    test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                    "stash[%ju] deviation - memblock[%ju].element_bound: expected=%u, observed=%u",
-                    i, j, shadow_mbll->element_bound, mbll->element_bound
-                );
-                return TACTYK_TESTSTATE__FAIL;
-            }
-            if (mbll->memblock_index != shadow_mbll->memblock_index) {
-                snprintf(
-                    test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                    "stash[%ju] deviation - memblock[%ju].memblock_index: expected=%u, observed=%u",
-                    i, j, shadow_mbll->memblock_index, mbll->memblock_index
-                );
-                return TACTYK_TESTSTATE__FAIL;
-            }
-            if (mbll->offset != shadow_mbll->offset) {
-                snprintf(
-                    test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                    "stash[%ju] deviation - memblock[%ju].offset: expected=%u, observed=%u",
-                    i, j, shadow_mbll->offset, mbll->offset
-                );
-                return TACTYK_TESTSTATE__FAIL;
-            }
-        }
-
-        #define STASH_CHK(PROP) \
-        if (shadow_stash->PROP != stash->PROP) { \
-            snprintf( \
-                test_state->report, TACTYK_TEST__REPORT_BUFSIZE, \
-                "stash[%ju] deviation - " #PROP ": expected=%jd, observed=%jd", \
-                i, shadow_stash->PROP, stash->PROP \
-            ); \
-            return TACTYK_TESTSTATE__FAIL; \
-        }
-        STASH_CHK(a.i64[0]);
-        STASH_CHK(a.i64[1]);
-        STASH_CHK(b.i64[0]);
-        STASH_CHK(b.i64[1]);
-        STASH_CHK(c.i64[0]);
-        STASH_CHK(c.i64[1]);
-        STASH_CHK(d.i64[0]);
-        STASH_CHK(d.i64[1]);
-        STASH_CHK(e.i64[0]);
-        STASH_CHK(e.i64[1]);
-
-        STASH_CHK(f.i64[0]);
-        STASH_CHK(f.i64[1]);
-        STASH_CHK(g.i64[0]);
-        STASH_CHK(g.i64[1]);
-        STASH_CHK(h.i64[0]);
-        STASH_CHK(h.i64[1]);
-        STASH_CHK(i.i64[0]);
-        STASH_CHK(i.i64[1]);
-        STASH_CHK(j.i64[0]);
-        STASH_CHK(j.i64[1]);
-
-        STASH_CHK(k.i64[0]);
-        STASH_CHK(k.i64[1]);
-        STASH_CHK(l.i64[0]);
-        STASH_CHK(l.i64[1]);
-        STASH_CHK(m.i64[0]);
-        STASH_CHK(m.i64[1]);
-        STASH_CHK(n.i64[0]);
-        STASH_CHK(n.i64[1]);
-        STASH_CHK(o.i64[0]);
-        STASH_CHK(o.i64[1]);
-
-        STASH_CHK(p.i64[0]);
-        STASH_CHK(p.i64[1]);
-        STASH_CHK(q.i64[0]);
-        STASH_CHK(q.i64[1]);
-        STASH_CHK(r.i64[0]);
-        STASH_CHK(r.i64[1]);
-        STASH_CHK(s.i64[0]);
-        STASH_CHK(s.i64[1]);
-        STASH_CHK(t.i64[0]);
-        STASH_CHK(t.i64[1]);
-
-        STASH_CHK(u.i64[0]);
-        STASH_CHK(u.i64[1]);
-        STASH_CHK(v.i64[0]);
-        STASH_CHK(v.i64[1]);
-        STASH_CHK(w.i64[0]);
-        STASH_CHK(w.i64[1]);
-        STASH_CHK(x.i64[0]);
-        STASH_CHK(x.i64[1]);
-        STASH_CHK(y.i64[0]);
-        STASH_CHK(y.i64[1]);
-
-        STASH_CHK(z.i64[0]);
-        STASH_CHK(z.i64[1]);
-        #undef STASH_CHK
-    }
-
-    return TACTYK_TESTSTATE__PASS;
-}
-
-uint64_t tactyk_test__CONTINUE(struct tactyk_dblock__DBlock *spec) {
-    return TACTYK_TESTSTATE__CONTINUE;
-}
-
-uint64_t tactyk_test__RETURN(struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *retval = spec->token->next;
-    ccall_retval = 0;
-    if (retval != NULL) {
-        tactyk_dblock__try_parseint(&ccall_retval, retval);
-    }
-    return TACTYK_TESTSTATE__PASS;
-}
-uint64_t tactyk_test__STATE(struct tactyk_dblock__DBlock *spec) {
-    if ( vmctx == NULL) {
-        tactyk_test__report("No asmvm context");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    struct tactyk_dblock__DBlock *td = spec->child;
-    while (td != NULL) {
-        struct tactyk_dblock__DBlock *item_name = td->token;
-        assert(item_name != NULL);
-        struct tactyk_dblock__DBlock *item_value = td->token->next;
-        if (item_value == NULL) {
-            tactyk_test__report("Unspecified item value");
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-
-        struct tactyk_test_entry *test = tactyk_dblock__get(base_tests, item_name);
-
-        if (test == NULL) {
-            char buf[256];
-            sprintf(buf, "No handler for test item: ");
-            tactyk_dblock__export_cstring(&buf[strlen(buf)], 256-strlen(buf), item_name );
-            tactyk_test__report(buf);
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-
-        if (test->adjust == NULL) {
-            tactyk_test__report("state adjustment function is undefined");
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-        if (!test->adjust(test, td)) {
-            tactyk_test__report("state adjustment rejected or not implemented");
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-
-        td = td->next;
-    }
-    return TACTYK_TESTSTATE__PASS;
-}
-uint64_t tactyk_test__DATA(struct tactyk_dblock__DBlock *spec) {
-    //struct tactyk_dblock__DBlock *mem_target = spec->token->next;
-    return TACTYK_TESTSTATE__PASS;
-}
-uint64_t tactyk_test__REF(struct tactyk_dblock__DBlock *spec) {
-    //struct tactyk_dblock__DBlock *mem_target = spec->token->next;
-    return TACTYK_TESTSTATE__PASS;
-}
-
-// If an "ERROR" command is executed, it means the system has not recognized an expected error condition, so this should always fail.
-//  (If the system does generate an error, that gets captured by the testing framework error handler and recognized by looking ahead)
-uint64_t tactyk_test__ERROR(struct tactyk_dblock__DBlock *spec) {
-    return TACTYK_TESTSTATE__FAIL;
-}
-
-bool tactyk_test__SET_CONTEXT_STATUS(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *value = spec->token->next;
-    int64_t ival = 0;
-    if (!tactyk_dblock__try_parseint(&ival, value)) {
-        tactyk_test__report("Test value parameter is not an integer");
-        return false;
-    }
-    vmctx->STATUS= (uint64_t) ival;
-    shadow_vmctx->STATUS = (uint64_t) ival;
-    return true;
-}
-uint64_t tactyk_test__TEST_CONTEXT_STATUS(struct tactyk_test_entry *valtest_spec, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *expected_value = spec->token->next;
-    int64_t ival = 0;
-    if (!tactyk_dblock__try_parseint(&ival, expected_value)) {
-        tactyk_test__report("Test value parameter is not an integer");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    shadow_vmctx->STATUS = vmctx->STATUS;
-    if (vmctx->STATUS == (uint64_t) ival) {
-        return TACTYK_TESTSTATE__PASS;
-    }
-    else {
-        sprintf(test_state->report, "context status deviation, expected:%ju observed:%ju", (uint64_t)ival, vmctx->STATUS);
-        return TACTYK_TESTSTATE__FAIL;
-    }
-}
-
-uint64_t tactyk_test__TEST_CALLBACK(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *expected_cbid = spec->token->next;
-    uint64_t uival = 0;
-    if (!tactyk_dblock__try_parseuint(&uival, expected_cbid)) {
-        tactyk_test__report("Callback-id parameter is not an integer");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    if (uival > 8) {
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "Invalid callback id: %ju", uival);
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    if (uival == callback_id) {
-        callback_id = 0;        // The callback has been properly accounted for at this point.
-        return TACTYK_TESTSTATE__PASS;
-    }
-    else {
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "Incorrect callback -- expected id: %ju, observed id:%ju", uival, callback_id);
-        return TACTYK_TESTSTATE__FAIL;
-    }
-}
-
-uint64_t read_spec__binary_data(struct tactyk_dblock__DBlock **out, struct tactyk_dblock__DBlock *spec) {
-
-    struct tactyk_dblock__DBlock *data_array = tactyk_dblock__new(0);
-    *out = data_array;
-
-    struct tactyk_dblock__DBlock *token = NULL;
-
-    while (spec != NULL) {
-        token = spec->token;
-        if (tactyk_dblock__equals_c_string(token, "byte")) {
-            token = token->next;
-            while (token != NULL) {
-                int64_t val = 0;
-                if (!tactyk_dblock__try_parseint(&val, token)) {
-                    goto parse_fail;
-                }
-                tactyk_dblock__append_byte(data_array, (uint8_t)val);
-                token = token->next;
-            }
-        }
-        else if (tactyk_dblock__equals_c_string(token, "word")) {
-            token = token->next;
-            while (token != NULL) {
-                int64_t val = 0;
-                if (!tactyk_dblock__try_parseint(&val, token)) {
-                    goto parse_fail;
-                }
-                tactyk_dblock__append_word(data_array, (uint16_t)val);
-                token = token->next;
-            }
-        }
-        else if (tactyk_dblock__equals_c_string(token, "dword")) {
-            token = token->next;
-            while (token != NULL) {
-                int64_t val = 0;
-                if (!tactyk_dblock__try_parseint(&val, token)) {
-                    goto parse_fail;
-                }
-                tactyk_dblock__append_dword(data_array, (uint32_t)val);
-                token = token->next;
-            }
-        }
-        else if (tactyk_dblock__equals_c_string(token, "qword")) {
-            token = token->next;
-            while (token != NULL) {
-                int64_t val = 0;
-                if (!tactyk_dblock__try_parseint(&val, token)) {
-                    goto parse_fail;
-                }
-                tactyk_dblock__append_qword(data_array, (uint64_t)val);
-                token = token->next;
-            }
-        }
-        else {
-            while (token != NULL) {
-                int64_t val = 0;
-                if (!tactyk_dblock__try_parseint(&val, token)) {
-                    goto parse_fail;
-                }
-                tactyk_dblock__append_qword(data_array, (uint64_t)val);
-                token = token->next;
-            }
-        }
-        spec = spec->next;
-    }
-    return TACTYK_TESTSTATE__PASS;
-
-    parse_fail: {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, token);
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "'%s' is not an integer", buf);
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-}
-
-uint64_t tactyk_test__TEST_MEM(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *name = spec->token->next;
-
-    // if the name is the name of a loaded program, then a "foreign" memblock is specified, so use the "foreign" program's memblock table.
-    //  Also:  I should probably generate a test errror for programs which share the same name as a memblock.
-    struct tactyk_asmvm__Program *src_program = tactyk_dblock__get(programs, name);
-    struct tactyk_asmvm__memblock_lowlevel *shadow_mbs = NULL;
-    if (src_program == NULL) {
-        src_program = tprg;
-        shadow_mbs = shadow_memblocks;
-    }
-    else {
-        shadow_mbs = tactyk_dblock__get(shadow_memblock_sets, name);
-        name = name->next;
-    }
-
-    struct tactyk_asmvm__memblock_highlevel *mbhl = tactyk_dblock__get(src_program->memory_layout_hl, name);
-    struct tactyk_asmvm__memblock_lowlevel *mbll = tactyk_dblock__index(src_program->memory_layout_ll, mbhl->memblock_id);
-    struct tactyk_asmvm__memblock_lowlevel *shadow_mbll = &shadow_mbs[mbhl->memblock_id];
-
-    uint64_t len = mbll->array_bound + mbll->element_bound -1;
-    uint64_t slen = shadow_mbll->array_bound + shadow_mbll->element_bound -1;
-    if (len != slen) {
-        snprintf(
-            test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-            "memblock unexpectedly replaced or resized -- expected length:%ju, observed length:%ju",
-            len, slen
-        );
-        return TACTYK_TESTSTATE__FAIL;
-    }
-
-    struct tactyk_dblock__DBlock *iindex = name->next;
-    if (iindex != NULL) {
-        if (tactyk_dblock__equals_c_string(iindex, "*")) {
-            // arbitrarilly accept the entire memblock
-            for (uint64_t i = 0; i < len; i += 1) {
-                shadow_mbll->base_address[i] = mbll->base_address[i];
-            }
-        }
-        else {
-            uint64_t idx = 0;
-            if (!tactyk_dblock__try_parseuint(&idx, iindex)) {
-                char buf[64];
-                tactyk_dblock__export_cstring(buf, 64, iindex);
-                snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "'%s' is not an integer", buf);
-                return TACTYK_TESTSTATE__TEST_ERROR;
-            }
-            struct tactyk_dblock__DBlock *item_type = iindex->next;
-            struct tactyk_dblock__DBlock *expected_value = item_type->next;
-            if (expected_value == NULL) {
-                expected_value = item_type;
-                item_type = NULL;
-            }
-            int64_t ival = 0;
-            if (!tactyk_dblock__try_parseint(&ival, expected_value)) {
-                char buf[64];
-                tactyk_dblock__export_cstring(buf, 64, expected_value);
-                snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "'%s' is not an integer", buf);
-                return TACTYK_TESTSTATE__TEST_ERROR;
-            }
-            if ( (item_type == NULL) || tactyk_dblock__equals_c_string(item_type, "byte") ) {
-                if (mbll->base_address[idx] != (uint8_t)ival) {
-                    char buf[64];
-                    tactyk_dblock__export_cstring(buf, 64, name);
-                    snprintf(
-                        test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                        "deviation in memblock '%s' at offset %ju - expected value:%u, observed value:%u",
-                        buf, idx, (uint8_t)ival, mbll->base_address[idx]
-                    );
-                    return TACTYK_TESTSTATE__FAIL;
-                }
-                shadow_mbll->base_address[idx] = (uint8_t)ival;
-            }
-            else if ( tactyk_dblock__equals_c_string(item_type, "word") ) {
-                if (*((uint16_t*) &mbll->base_address[idx]) != (uint16_t)ival) {
-                    char buf[64];
-                    tactyk_dblock__export_cstring(buf, 64, name);
-                    snprintf(
-                        test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                        "deviation in memblock '%s' at offset %ju - expected value:%u, observed value:%u",
-                        buf, idx, (uint16_t)ival, *((uint16_t*) &mbll->base_address[idx])
-                    );
-                    return TACTYK_TESTSTATE__FAIL;
-                }
-               *((uint16_t*) &shadow_mbll->base_address[idx]) = (uint16_t)ival;
-            }
-
-            else if ( tactyk_dblock__equals_c_string(item_type, "dword") ) {
-                if (*((uint32_t*) &mbll->base_address[idx]) != (uint32_t)ival) {
-                    char buf[64];
-                    tactyk_dblock__export_cstring(buf, 64, name);
-                    snprintf(
-                        test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                        "deviation in memblock '%s' at offset %ju - expected value:%u, observed value:%u",
-                        buf, idx, (uint32_t)ival, *((uint32_t*) &mbll->base_address[idx])
-                    );
-                    return TACTYK_TESTSTATE__FAIL;
-                }
-               *((uint32_t*) &shadow_mbll->base_address[idx]) = (uint32_t)ival;
-            }
-            else if ( tactyk_dblock__equals_c_string(item_type, "qword") ) {
-                if (*((uint64_t*) &mbll->base_address[idx]) != (uint64_t)ival) {
-                    char buf[64];
-                    tactyk_dblock__export_cstring(buf, 64, name);
-                    snprintf(
-                        test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                        "deviation in memblock '%s' at offset %ju - expected value:%ju, observed value:%ju",
-                        buf, idx, (uint64_t)ival, *((uint64_t*) &mbll->base_address[idx])
-                    );
-                    return TACTYK_TESTSTATE__FAIL;
-                }
-               *((uint64_t*) &shadow_mbll->base_address[idx]) = (uint64_t)ival;
-            }
-            else {
-                char buf[64];
-                tactyk_dblock__export_cstring(buf, 64, item_type);
-                snprintf(
-                    test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                    "Unrecognized data type: %s",
-                    buf
-                );
-
-                return TACTYK_TESTSTATE__TEST_ERROR;
-            }
-
-        }
-    }
-
-
-    struct tactyk_dblock__DBlock *mspec = spec->child;
-    if (mspec != NULL) {
-        // actually test the entire memblock
-    }
-
-    return TACTYK_TESTSTATE__PASS;
-}
-
-bool tactyk_test__SET_MEM(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *value = spec->token->next;
-    return false;
-}
-
-uint64_t tactyk_test__ALLOC(struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *name = spec->token->next;
-    struct tactyk_dblock__DBlock *data;
-
-    if (read_spec__binary_data(&data, spec->child) == TACTYK_TESTSTATE__TEST_ERROR) {
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-
-    struct tactyk_asmvm__memblock_highlevel *mhl = tactyk_dblock__get(tprg->memory_layout_hl, name);
-    uint64_t idx = mhl->memblock_id;
-    struct tactyk_asmvm__memblock_lowlevel *mll = &vmctx->memblocks[idx];
-
-    uint64_t len = data->length;
-    uint8_t *bytes = tactyk_dblock__release(data);
-    uint8_t *shadow_bytes = calloc(1, len);
-    memcpy(shadow_bytes, bytes, len);
-
-    mhl->data = bytes;
-    mll->base_address = bytes;
-    mll->array_bound = 1;
-    mll->element_bound = len;
-
-    struct tactyk_asmvm__memblock_lowlevel *shadow_mll = &shadow_memblocks[idx];
-    shadow_mll->base_address = shadow_bytes;
-    shadow_mll->array_bound = 1;
-    shadow_mll->element_bound = len;
-    //mll = mhl->memblock;
-
-    return TACTYK_TESTSTATE__PASS;
-}
-
-
-bool tactyk_test__SET_ADDR (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *value = spec->token->next;
-
-    if (tactyk_dblock__count_peers(value) < 2) {
-        tactyk_test__report("Not enough arguments to specify a memory binding.");
-        return false;
-    }
-
-    struct tactyk_dblock__DBlock *name = value;
-    struct tactyk_dblock__DBlock *reg_ofs = value->next;
-
-    struct tactyk_asmvm__memblock_highlevel *mbhl = tactyk_dblock__get(vmctx->hl_program_ref->memory_layout_hl, name);
-    struct tactyk_asmvm__memblock_lowlevel *mbll = &vmctx->memblocks[mbhl->memblock_id];
-    struct tactyk_asmvm__memblock_lowlevel *shadow_mbll = &shadow_memblocks[mbhl->memblock_id];
-
-    struct tactyk_asmvm__memblock_lowlevel *target = NULL;
-    struct tactyk_asmvm__memblock_lowlevel *shadow_target = NULL;
-
-    int64_t ofs = 0;
-    uint64_t abound = mbll->array_bound;
-    uint64_t ebound = mbll->element_bound;
-    if (!tactyk_dblock__try_parseint(&ofs, reg_ofs)) {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, reg_ofs);
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "address-offset '%s' is not an integer.", buf);
-        return false;
-    }
-    switch(entry->element_offset) {
-        case 1: {
-            vmctx->reg.rADDR1 = (uint64_t*) &mbll->base_address[ofs];
-            shadow_vmctx->reg.rADDR1 = (uint64_t*) &shadow_mbll->base_address[ofs];
-            break;
-        }
-        case 2: {
-            vmctx->reg.rADDR2 = (uint64_t*) &mbll->base_address[ofs];
-            shadow_vmctx->reg.rADDR2 = (uint64_t*) &shadow_mbll->base_address[ofs];
-            break;
-        }
-        case 3: {
-            vmctx->reg.rADDR3 = (uint64_t*) &mbll->base_address[ofs];
-            shadow_vmctx->reg.rADDR3 = (uint64_t*) &shadow_mbll->base_address[ofs];
-            break;
-        }
-        case 4: {
-            vmctx->reg.rADDR4 = (uint64_t*) &mbll->base_address[ofs];
-            shadow_vmctx->reg.rADDR4 = (uint64_t*) &shadow_mbll->base_address[ofs];
-            break;
-        }
-    }
-
-    target = &vmctx->memblocks[entry->element_offset];
-    target->base_address = mbll->base_address;
-    target->array_bound = abound;
-    target->element_bound = ebound;
-
-    shadow_target = &shadow_memblocks[entry->element_offset];
-    shadow_target->base_address = shadow_mbll->base_address;
-    shadow_target->array_bound = abound;
-    shadow_target->element_bound = ebound;
-
-    return true;
-}
-
-uint64_t tactyk_test__TEST_ADDR(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *expected_value = spec->token->next;
-
-    uint64_t num_tokens = tactyk_dblock__count_peers(expected_value);
-    if (num_tokens < 2) {
-        tactyk_test__report("Not enough arguments to specify a memory binding.");
-        return false;
-    }
-    struct tactyk_dblock__DBlock *name = expected_value;
-    struct tactyk_dblock__DBlock *reg_ofs = expected_value->next;
-
-    // if the name is the name of a loaded program, then a "foreign" memblock is specified, so use the "foreign" program's memblock table.
-    //  Also:  I should probably generate a test errror for programs which share the same name as a memblock.
-    struct tactyk_asmvm__Program *src_program = tactyk_dblock__get(programs, name);
-    struct tactyk_asmvm__memblock_lowlevel *shadow_mbs = NULL;
-    if (src_program == NULL) {
-        src_program = tprg;
-        shadow_mbs = shadow_memblocks;
-    }
-    else {
-        shadow_mbs = tactyk_dblock__get(shadow_memblock_sets, name);
-        reg_ofs = reg_ofs->next;
-        name = name->next;
-    }
-
-    struct tactyk_asmvm__memblock_highlevel *mbhl = NULL;
-    struct tactyk_asmvm__memblock_lowlevel *mbll = NULL;
-    struct tactyk_asmvm__memblock_lowlevel *shadow_mbll = NULL;
-
-    uint64_t mb_index = 0;
-    if (tactyk_dblock__try_parseuint(&mb_index, name)) {
-        mbhl = tactyk_dblock__index_allocated(src_program->memory_layout_hl->store, mb_index);
-        mbll = tactyk_dblock__index_allocated(src_program->memory_layout_ll, mb_index);
-        shadow_mbll = &shadow_mbs[mb_index];
-    }
-    else {
-        mbhl = tactyk_dblock__get(src_program->memory_layout_hl, name);
-        mbll = tactyk_dblock__index(src_program->memory_layout_ll, mbhl->memblock_id);
-        shadow_mbll = &shadow_mbs[mbhl->memblock_id];
-    }
-
-    struct tactyk_asmvm__memblock_lowlevel *target = &vmctx->active_memblocks[entry->element_offset-1];
-    struct tactyk_asmvm__memblock_lowlevel *shadow_target = &shadow_vmctx->active_memblocks[entry->element_offset-1];
-    int64_t expected_ofs = 0;
-    uint64_t expected_abound = 0;
-    uint64_t expected_ebound = 0;
-
-    if (!tactyk_dblock__try_parseint(&expected_ofs, reg_ofs)) {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, reg_ofs);
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "address-offset '%s' is not an integer.", buf);
-        return false;
-    }
-    if (num_tokens >= 4) {
-        struct tactyk_dblock__DBlock *array_bound = reg_ofs->next;
-        struct tactyk_dblock__DBlock *element_bound = reg_ofs->next->next;
-        if (!tactyk_dblock__try_parseuint(&expected_abound, array_bound)) {
-            char buf[64];
-            tactyk_dblock__export_cstring(buf, 64, array_bound);
-            snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "array-bound '%s' is not an integer.", buf);
-            return false;
-        }
-        if (!tactyk_dblock__try_parseuint(&expected_ebound, element_bound)) {
-            char buf[64];
-            tactyk_dblock__export_cstring(buf, 64, element_bound);
-            snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "element-bound '%s' is not an integer.", buf);
-            return false;
-        }
-    }
-    else {
-        expected_abound = mbll->array_bound;
-        expected_ebound = mbll->element_bound;
-    }
-
-    if ( target->memblock_index != mbll->memblock_index ) {
-        snprintf(
-            test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-            "memory binding deviation -- expected memblock id=%u observed memblock id=%u",
-            mbll->memblock_index, target->memblock_index
-        );
-        return TACTYK_TESTSTATE__FAIL;
-    }
-    if ( target->base_address != mbll->base_address ) {
-        snprintf(
-            test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-            "memory binding deviation -- expected memblock ptr=%p observed memblock ptr=%p",
-            mbll->base_address, target->base_address
-        );
-        return TACTYK_TESTSTATE__FAIL;
-    }
-    if ( target->array_bound != expected_abound ) {
-        snprintf(
-            test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-            "memory binding deviation -- expected memblock array-bound=%ju observed memblock array-bound=%u",
-            expected_abound, target->array_bound
-        );
-        return TACTYK_TESTSTATE__FAIL;
-    }
-    if ( target->element_bound != expected_ebound ) {
-        snprintf(
-            test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-            "memory binding deviation -- expected memblock element-bound=%ju observed memblock element-bound=%u",
-            expected_ebound, target->element_bound
-        );
-        return TACTYK_TESTSTATE__FAIL;
-    }
-    if ( target->offset != mbll->offset ) {
-        snprintf(
-            test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-            "memory binding deviation -- expected memblock offset=%u observed memblock offset=%u",
-            mbll->offset, target->offset
-        );
-        return TACTYK_TESTSTATE__FAIL;
-    }
-
-    int64_t observed_ofs = 0;
-
-    switch(entry->element_offset) {
-        case 1: {
-            observed_ofs = (uint8_t*) vmctx->reg.rADDR1 - (uint8_t*)target->base_address;
-            shadow_vmctx->reg.rADDR1 = (uint64_t*) &target->base_address[expected_ofs];
-            break;
-        }
-        case 2: {
-            observed_ofs = (uint8_t*) vmctx->reg.rADDR2 - (uint8_t*)target->base_address;
-            shadow_vmctx->reg.rADDR2 = (uint64_t*) &target->base_address[expected_ofs];
-            break;
-        }
-        case 3: {
-            observed_ofs = (uint8_t*) vmctx->reg.rADDR3 - (uint8_t*)target->base_address;
-            shadow_vmctx->reg.rADDR3 = (uint64_t*) &target->base_address[expected_ofs];
-            break;
-        }
-        case 4: {
-            observed_ofs = (uint8_t*) vmctx->reg.rADDR4 - (uint8_t*)target->base_address;
-            shadow_vmctx->reg.rADDR4 = (uint64_t*) &target->base_address[expected_ofs];
-            break;
-        }
-    }
-
-    shadow_target->base_address = target->base_address;
-    shadow_target->array_bound = expected_abound;
-    shadow_target->element_bound = expected_ebound;
-    shadow_target->memblock_index = target->memblock_index;
-    shadow_target->offset = target->offset;
-
-    if (observed_ofs != expected_ofs) {
-        snprintf(
-            test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-            "memory binding deviation -- expected memblock offset=%jd observed memblock offset=%jd",
-            expected_ofs, observed_ofs
-        );
-        return TACTYK_TESTSTATE__FAIL;
-    }
-
-    return TACTYK_TESTSTATE__PASS;
-}
-
-bool tactyk_test__SET_REGISTER (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *value = spec->token->next;
-
-    int64_t ival = 0;
-    if (!tactyk_dblock__try_parseint(&ival, value)) {
-        tactyk_test__report("Test value parameter is not an integer");
-        return false;
-    }
-    switch(entry->element_offset) {
-        case 1: {
-            vmctx->reg.rLWCSI = (uint64_t) ival;
-            shadow_vmctx->reg.rLWCSI = (uint64_t) ival;
-            return true;
-        }
-        case 2: {
-            vmctx->reg.rMCSI = (uint64_t) ival;
-            shadow_vmctx->reg.rMCSI = (uint64_t) ival;
-            return true;
-        }
-        case 10: {
-            vmctx->reg.rA = (uint64_t) ival;
-            shadow_vmctx->reg.rA = (uint64_t) ival;
-            return true;
-        }
-        case 11: {
-            vmctx->reg.rB = (uint64_t) ival;
-            shadow_vmctx->reg.rB = (uint64_t) ival;
-            return true;
-        }
-        case 12: {
-            vmctx->reg.rC = (uint64_t) ival;
-            shadow_vmctx->reg.rC = (uint64_t) ival;
-            return true;
-        }
-        case 13: {
-            vmctx->reg.rD = (uint64_t) ival;
-            shadow_vmctx->reg.rD = (uint64_t) ival;
-            return true;
-        }
-        case 14: {
-            vmctx->reg.rE = (uint64_t) ival;
-            shadow_vmctx->reg.rE = (uint64_t) ival;
-            return true;
-        }
-        case 15: {
-            vmctx->reg.rF = (uint64_t) ival;
-            shadow_vmctx->reg.rF = (uint64_t) ival;
-            return true;
-        }
-        default: {
-            tactyk_test__report("Test element-offset is invalid");
-            return false;
-        }
-    }
-}
-
-uint64_t tactyk_test__TEST_REGISTER(struct tactyk_test_entry *valtest_spec, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *expected_value = spec->token->next;
-
-    int64_t ival = 0;
-    if (!tactyk_dblock__try_parseint(&ival, expected_value)) {
-        tactyk_test__report("Test value parameter is not an integer");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    int64_t stval = 0;
-    switch(valtest_spec->element_offset) {
-        case 1: {
-            stval = vmctx->reg.rLWCSI;
-            shadow_vmctx->reg.rLWCSI = vmctx->reg.rLWCSI;
-            break;
-        }
-        case 2: {
-            stval = vmctx->reg.rMCSI;
-            shadow_vmctx->reg.rMCSI = vmctx->reg.rMCSI;
-            break;
-        }
-        case 10: {
-            stval = vmctx->reg.rA;
-            shadow_vmctx->reg.rA = vmctx->reg.rA;
-            break;
-        }
-        case 11: {
-            stval = vmctx->reg.rB;
-            shadow_vmctx->reg.rB = vmctx->reg.rB;
-            break;
-        }
-        case 12: {
-            stval = vmctx->reg.rC;
-            shadow_vmctx->reg.rC = vmctx->reg.rC;
-            break;
-        }
-        case 13: {
-            stval = vmctx->reg.rD;
-            shadow_vmctx->reg.rD = vmctx->reg.rD;
-            break;
-        }
-        case 14: {
-            stval = vmctx->reg.rE;
-            shadow_vmctx->reg.rE = vmctx->reg.rE;
-            break;
-        }
-        case 15: {
-            stval = vmctx->reg.rF;
-            shadow_vmctx->reg.rF = vmctx->reg.rF;
-            break;
-        }
-        default: {
-            tactyk_test__report("Test element-offset is invalid");
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-    }
-
-    if (stval == ival) {
-        return TACTYK_TESTSTATE__PASS;
-    }
-    else {
-        sprintf(test_state->report, "deviation on register %s, expected:%jd observed:%jd", valtest_spec->name, ival, stval);
-        return TACTYK_TESTSTATE__FAIL;
-    }
-}
-
-bool tactyk_test__SET_XMM_REGISTER_FLOAT (struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *value = spec->token->next;
-
-    double fval = 0;
-    if (!tactyk_dblock__try_parsedouble(&fval, value)) {
-        tactyk_test__report("Test parameter is not a floating point number");
-        return false;
-    }
-    switch(entry->element_offset) {
-        case 0: {
-            vmctx->reg.xa.f64[0] = fval;
-            shadow_vmctx->reg.xa.f64[0] = fval;
-            break;
-        }
-        case 1: {
-            vmctx->reg.xb.f64[0] = fval;
-            shadow_vmctx->reg.xb.f64[0] = fval;
-            break;
-        }
-        case 2: {
-            vmctx->reg.xc.f64[0] = fval;
-            shadow_vmctx->reg.xc.f64[0] = fval;
-            break;
-        }
-        case 3: {
-            vmctx->reg.xd.f64[0] = fval;
-            shadow_vmctx->reg.xd.f64[0] = fval;
-            break;
-        }
-        case 4: {
-            vmctx->reg.xe.f64[0] = fval;
-            shadow_vmctx->reg.xe.f64[0] = fval;
-            break;
-        }
-        case 5: {
-            vmctx->reg.xf.f64[0] = fval;
-            shadow_vmctx->reg.xf.f64[0] = fval;
-            break;
-        }
-        case 6: {
-            vmctx->reg.xg.f64[0] = fval;
-            shadow_vmctx->reg.xg.f64[0] = fval;
-            break;
-        }
-        case 7: {
-            vmctx->reg.xh.f64[0] = fval;
-            shadow_vmctx->reg.xh.f64[0] = fval;
-            break;
-        }
-        case 8: {
-            vmctx->reg.xi.f64[0] = fval;
-            shadow_vmctx->reg.xi.f64[0] = fval;
-            break;
-        }
-        case 9: {
-            vmctx->reg.xj.f64[0] = fval;
-            shadow_vmctx->reg.xj.f64[0] = fval;
-            break;
-        }
-        case 10: {
-            vmctx->reg.xk.f64[0] = fval;
-            shadow_vmctx->reg.xk.f64[0] = fval;
-            break;
-        }
-        case 11: {
-            vmctx->reg.xl.f64[0] = fval;
-            shadow_vmctx->reg.xl.f64[0] = fval;
-            break;
-        }
-        case 12: {
-            vmctx->reg.xm.f64[0] = fval;
-            shadow_vmctx->reg.xm.f64[0] = fval;
-            break;
-        }
-        case 13: {
-            vmctx->reg.xn.f64[0] = fval;
-            shadow_vmctx->reg.xn.f64[0] = fval;
-            break;
-        }
-        case 14: {
-            vmctx->reg.xTEMPA.f64[0] = fval;
-            shadow_vmctx->reg.xTEMPA.f64[0] = fval;
-            break;
-        }
-        case 15: {
-            vmctx->reg.xTEMPB.f64[0] = fval;
-            shadow_vmctx->reg.xTEMPB.f64[0] = fval;
-            break;
-        }
-        default: {
-            tactyk_test__report("Test element-offset is invalid");
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-    }
-    return TACTYK_TESTSTATE__PASS;
-}
-
-uint64_t tactyk_test__TEST_XMM_REGISTER_FLOAT (struct tactyk_test_entry *valtest_spec, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *expected_value = spec->token->next;
-
-    double fval = 0;
-    if (!tactyk_dblock__try_parsedouble(&fval, expected_value)) {
-        tactyk_test__report("Test parameter is not a floating point number");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    double stval = 0;
-    switch(valtest_spec->element_offset) {
-        case 0: {
-            stval = vmctx->reg.xa.f64[0];
-            shadow_vmctx->reg.xa.f64[0] = vmctx->reg.xa.f64[0];
-            break;
-        }
-        case 1: {
-            stval = vmctx->reg.xb.f64[0];
-            shadow_vmctx->reg.xb.f64[0] = vmctx->reg.xb.f64[0];
-            break;
-        }
-        case 2: {
-            stval = vmctx->reg.xc.f64[0];
-            shadow_vmctx->reg.xc.f64[0] = vmctx->reg.xc.f64[0];
-            break;
-        }
-        case 3: {
-            stval = vmctx->reg.xd.f64[0];
-            shadow_vmctx->reg.xd.f64[0] = vmctx->reg.xd.f64[0];
-            break;
-        }
-        case 4: {
-            stval = vmctx->reg.xe.f64[0];
-            shadow_vmctx->reg.xe.f64[0] = vmctx->reg.xe.f64[0];
-            break;
-        }
-        case 5: {
-            stval = vmctx->reg.xf.f64[0];
-            shadow_vmctx->reg.xf.f64[0] = vmctx->reg.xf.f64[0];
-            break;
-        }
-        case 6: {
-            stval = vmctx->reg.xg.f64[0];
-            shadow_vmctx->reg.xg.f64[0] = vmctx->reg.xg.f64[0];
-            break;
-        }
-        case 7: {
-            stval = vmctx->reg.xh.f64[0];
-            shadow_vmctx->reg.xh.f64[0] = vmctx->reg.xh.f64[0];
-            break;
-        }
-        case 8: {
-            stval = vmctx->reg.xi.f64[0];
-            shadow_vmctx->reg.xi.f64[0] = vmctx->reg.xi.f64[0];
-            break;
-        }
-        case 9: {
-            stval = vmctx->reg.xj.f64[0];
-            shadow_vmctx->reg.xj.f64[0] = vmctx->reg.xj.f64[0];
-            break;
-        }
-        case 10: {
-            stval = vmctx->reg.xk.f64[0];
-            shadow_vmctx->reg.xk.f64[0] = vmctx->reg.xk.f64[0];
-            break;
-        }
-        case 11: {
-            stval = vmctx->reg.xl.f64[0];
-            shadow_vmctx->reg.xl.f64[0] = vmctx->reg.xl.f64[0];
-            break;
-        }
-        case 12: {
-            stval = vmctx->reg.xm.f64[0];
-            shadow_vmctx->reg.xm.f64[0] = vmctx->reg.xm.f64[0];
-            break;
-        }
-        case 13: {
-            stval = vmctx->reg.xn.f64[0];
-            shadow_vmctx->reg.xn.f64[0] = vmctx->reg.xn.f64[0];
-            break;
-        }
-        case 14: {
-            stval = vmctx->reg.xTEMPA.f64[0];
-            shadow_vmctx->reg.xTEMPA.f64[0] = vmctx->reg.xTEMPA.f64[0];
-            break;
-        }
-        case 15: {
-            stval = vmctx->reg.xTEMPB.f64[0];
-            shadow_vmctx->reg.xTEMPB.f64[0] = vmctx->reg.xTEMPB.f64[0];
-            break;
-        }
-        default: {
-            tactyk_test__report("Test element-offset is invalid");
-            return TACTYK_TESTSTATE__TEST_ERROR;
-        }
-    }
-    stval = fabs(stval - fval);
-    if (stval <= precision) {
-        return TACTYK_TESTSTATE__PASS;
-    }
-    else {
-        sprintf(test_state->report, "deviation on register %s, expected:%f observed:%f", valtest_spec->name, fval, stval);
-        return TACTYK_TESTSTATE__FAIL;
-    }
-}
-
-uint64_t tactyk_test__TEST_LWCALL_STACK(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    uint64_t num_tokens = tactyk_dblock__count_tokens(spec);
-    if (num_tokens < 2) {
-        tactyk_test__report("[lwcstack] Not enough arguments");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    struct tactyk_dblock__DBlock *ofs_token = spec->token->next;
-    struct tactyk_dblock__DBlock *val_token = ofs_token->next;
-    struct tactyk_dblock__DBlock *rpt_token = val_token->next;
-
-    uint64_t ofs = 0;
-    uint64_t val = 0;
-    uint64_t rpt = 1;
-
-    if (!tactyk_dblock__try_parseuint(&ofs, ofs_token)) {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, ofs_token);
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "[lwcstack] Invalid offset: %s", buf);
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    if (!tactyk_dblock__try_parseuint(&val, val_token)) {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, val_token);
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "[lwcstack] Invalid stack index: %s", buf);
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    if ((rpt_token != NULL) && (!tactyk_dblock__try_parseuint(&rpt, rpt_token))) {
-        char buf[64];
-        tactyk_dblock__export_cstring(buf, 64, rpt_token);
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "[lwcstack] Invalid repetition qualifier: %s", buf);
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-
-    uint64_t limit = ofs + rpt;
-    if (limit > TACTYK_ASMVM__LWCALL_STACK_SIZE) {
-        limit = TACTYK_ASMVM__LWCALL_STACK_SIZE;
-    }
-
-    for (;ofs < limit; ofs++) {
-        if (vmctx->lwcall_stack[ofs] != val) {
-            snprintf(
-                test_state->report, TACTYK_TEST__REPORT_BUFSIZE,
-                "LW Call Stack deviation at offset %ju, expected=%ju observed=%u",
-                ofs, val, vmctx->lwcall_stack[ofs]
-            );
-            return TACTYK_TESTSTATE__FAIL;
-        }
-        shadow_lwcall_stack[ofs] = val;
-    }
-    return TACTYK_TESTSTATE__PASS;
-}
-
-uint64_t tactyk_test__TEST_STASH(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *ofs_token = spec->token->next;
-    struct tactyk_dblock__DBlock *fieldname_token = ofs_token->next;
-
-    uint64_t ofs = 0;
-    if (!tactyk_dblock__try_parseuint(&ofs, ofs_token)) {
-        ofs = vmctx->reg.rMCSI;
-        fieldname_token = ofs_token;
-    }
-    else if (ofs >= TACTYK_ASMVM__MCTX_STACK_SIZE) {
-        snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "mctx stack offset is out of bounds: %ju", ofs);
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-
-    struct tactyk_dblock__DBlock *val_token = fieldname_token->next;
-
-    struct tactyk_asmvm__MicrocontextStash *stash = &vmctx->microcontext_stack[ofs];
-    struct tactyk_asmvm__MicrocontextStash *shstash = &shadow_mctxstack[ofs];
-
-    char fn[64];
-    tactyk_dblock__export_cstring(fn, 64, fieldname_token);
-
-    double f64val = 0;
-    int64_t ival = 0;
-    double st_f64val = 0;
-    int64_t st_ival = 0;
-
-    bool pass = false;
-    bool field_matched = false;
-
-    tactyk_dblock__try_parsedouble(&f64val, val_token);
-    tactyk_dblock__try_parseint(&ival, val_token);
-
-    #define STASH_TEST(NAME, FIELD) \
-    else if (strncmp(fn, #NAME, 64) == 0) { \
-        pass = (ival == stash->FIELD); \
-        shstash->FIELD = ival; \
-        st_ival = stash->FIELD; \
-        field_matched = true; \
-    }
-    #define STASH_ATEST(NAME, FIELD, TYPE) \
-    else if (strncmp(fn, NAME, 64) == 0) { \
-        pass = ((TYPE)ival == stash->FIELD); \
-        shstash->FIELD = (TYPE)ival; \
-        st_ival = stash->FIELD; \
-        field_matched = true; \
-    }
-    if ( (strncmp(fn, "addr", 4) == 0) && (strlen(fn) == 5) ) {
-        uint64_t aofs = fn[4] - '1';
-        struct tactyk_asmvm__memblock_lowlevel *mbll = &stash->memblocks[aofs];
-        struct tactyk_asmvm__memblock_lowlevel *shmbll = &shstash->memblocks[aofs];
-        if (ival == 0) {
-            struct tactyk_dblock__DBlock *prgref = val_token;
-            struct tactyk_dblock__DBlock *addrref = val_token->next;
-            struct tactyk_asmvm__Program *prg = tactyk_dblock__get(programs, prgref);
-            if (prg == NULL) {
-                addrref = val_token;
-                prg = tprg;
-            }
-            struct tactyk_dblock__DBlock *addrofs = addrref->next;
-            struct tactyk_asmvm__memblock_highlevel *mbhl = tactyk_dblock__get(prg->memory_layout_hl, addrref);
-            if (mbhl == NULL) {
-                char buf[64];
-                tactyk_dblock__export_cstring(buf, 64, prgref);
-                snprintf(test_state->report, TACTYK_TEST__REPORT_BUFSIZE, "Undefined memblock:  '%s'", buf);
-                return TACTYK_TESTSTATE__TEST_ERROR;
-            }
-            ival += (int64_t)mbhl->data;
-
-            if (addrofs != NULL) {
-                uint64_t uival = 0;
-                tactyk_dblock__try_parseuint(&uival, addrofs);
-                ival += uival;
-            }
-        }
-        pass = ((uint8_t*)ival == mbll->base_address);
-        shmbll->base_address = (uint8_t*)ival;
-        field_matched = true;
-    }
-    STASH_ATEST("addr1.array_bound", memblocks[0].array_bound, uint32_t)
-    STASH_ATEST("addr1.element_bound", memblocks[0].element_bound, uint32_t)
-    STASH_ATEST("addr1.index", memblocks[0].memblock_index, uint32_t)
-    STASH_ATEST("addr1.offset", memblocks[0].offset, uint32_t)
-    STASH_ATEST("addr2.array_bound", memblocks[1].array_bound, uint32_t)
-    STASH_ATEST("addr2.element_bound", memblocks[1].element_bound, uint32_t)
-    STASH_ATEST("addr2.index", memblocks[1].memblock_index, uint32_t)
-    STASH_ATEST("addr2.offset", memblocks[1].offset, uint32_t)
-    STASH_ATEST("addr3.array_bound", memblocks[2].array_bound, uint32_t)
-    STASH_ATEST("addr3.element_bound", memblocks[2].element_bound, uint32_t)
-    STASH_ATEST("addr3.index", memblocks[2].memblock_index, uint32_t)
-    STASH_ATEST("addr3.offset", memblocks[2].offset, uint32_t)
-    STASH_ATEST("addr4.array_bound", memblocks[3].array_bound, uint32_t)
-    STASH_ATEST("addr4.element_bound", memblocks[3].element_bound, uint32_t)
-    STASH_ATEST("addr4.index", memblocks[3].memblock_index, uint32_t)
-    STASH_ATEST("addr4.offset", memblocks[3].offset, uint32_t)
-
-    STASH_TEST(al, a.i64[0])
-    STASH_TEST(ah, a.i64[1])
-    STASH_TEST(bl, b.i64[0])
-    STASH_TEST(bh, b.i64[1])
-    STASH_TEST(cl, c.i64[0])
-    STASH_TEST(ch, c.i64[1])
-    STASH_TEST(dl, d.i64[0])
-    STASH_TEST(dh, d.i64[1])
-    STASH_TEST(el, e.i64[0])
-    STASH_TEST(eh, e.i64[1])
-
-    STASH_TEST(fl, f.i64[0])
-    STASH_TEST(fh, f.i64[1])
-    STASH_TEST(gl, g.i64[0])
-    STASH_TEST(gh, g.i64[1])
-    STASH_TEST(hl, h.i64[0])
-    STASH_TEST(hh, h.i64[1])
-    STASH_TEST(il, i.i64[0])
-    STASH_TEST(ih, i.i64[1])
-    STASH_TEST(jl, j.i64[0])
-    STASH_TEST(jh, j.i64[1])
-
-    STASH_TEST(kl, k.i64[0])
-    STASH_TEST(kh, k.i64[1])
-    STASH_TEST(ll, l.i64[0])
-    STASH_TEST(lh, l.i64[1])
-    STASH_TEST(ml, m.i64[0])
-    STASH_TEST(mh, m.i64[1])
-    STASH_TEST(nl, n.i64[0])
-    STASH_TEST(nh, n.i64[1])
-    STASH_TEST(ol, o.i64[0])
-    STASH_TEST(oh, o.i64[1])
-
-    STASH_TEST(pl, p.i64[0])
-    STASH_TEST(ph, p.i64[1])
-    STASH_TEST(ql, q.i64[0])
-    STASH_TEST(qh, q.i64[1])
-    STASH_TEST(rl, r.i64[0])
-    STASH_TEST(rh, r.i64[1])
-    STASH_TEST(sl, s.i64[0])
-    STASH_TEST(sh, s.i64[1])
-    STASH_TEST(tl, t.i64[0])
-    STASH_TEST(th, t.i64[1])
-
-    STASH_TEST(ul, u.i64[0])
-    STASH_TEST(uh, u.i64[1])
-    STASH_TEST(vl, v.i64[0])
-    STASH_TEST(vh, v.i64[1])
-    STASH_TEST(wl, w.i64[0])
-    STASH_TEST(wh, w.i64[1])
-    STASH_TEST(xl, x.i64[0])
-    STASH_TEST(xh, x.i64[1])
-    STASH_TEST(yl, y.i64[0])
-    STASH_TEST(yh, y.i64[1])
-
-    STASH_TEST(zl, z.i64[0])
-    STASH_TEST(zh, z.i64[1])
-    #undef STASH_TEST
-    #undef STASH_ATEST
-
-    if (field_matched) {
-        if (pass) {
-            return TACTYK_TESTSTATE__PASS;
-        }
-        else {
-            sprintf(test_state->report, "deviation on mctx-stash field '%s', expected:%jd observed:%jd", fn, ival, st_ival);
-            return TACTYK_TESTSTATE__FAIL;
-        }
-    }
-    else {
-        sprintf(test_state->report, "unrecognized mctx field: '%s'", fn);
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-
-}
-
-uint64_t tactyk_test__TEST_CCALL_ARGUMENT(struct tactyk_test_entry *entry, struct tactyk_dblock__DBlock *spec) {
-    struct tactyk_dblock__DBlock *expected_value = spec->token->next;
-
-    int64_t ival = 0;
-    if (!tactyk_dblock__try_parseint(&ival, expected_value)) {
-        tactyk_test__report("Test value parameter is not an integer");
-        return TACTYK_TESTSTATE__TEST_ERROR;
-    }
-    assert(entry->element_offset < 6);
-    int64_t cbval = ccall_args[entry->element_offset];
-    if (ival == cbval) {
-        return TACTYK_TESTSTATE__PASS;
-    }
-    else {
-        sprintf(test_state->report, "callback argument #%ju deviation, expected:%jd observed:%jd", entry->element_offset, ival, cbval);
-        return TACTYK_TESTSTATE__FAIL;
-    }
-}
-
-void tactyk_test__mk_ccallarg_test(char *name, uint64_t ofs) {
-    struct tactyk_test_entry *entry = tactyk_dblock__new_managedobject(base_tests, name);
-    entry->name = name;
-    entry->adjust = NULL;
-    entry->test = tactyk_test__TEST_CCALL_ARGUMENT;
-    entry->element_offset = ofs;
-    entry->array_offset = 0;
-}
-
-void tactyk_test__mk_var_test(char *name, tactyk_test__VALUE_ADJUSTER setter, tactyk_test__VALUE_TESTER *tester) {
-    struct tactyk_test_entry *entry = tactyk_dblock__new_managedobject(base_tests, name);
-    entry->name = name;
-    entry->adjust = setter;
-    entry->test = tester;
-    entry->element_offset = 0;
-    entry->array_offset = 0;
-}
-
-void tactyk_test__mk_register_test(char *name, uint64_t ofs) {
-    struct tactyk_test_entry *entry = tactyk_dblock__new_managedobject(base_tests, name);
-    entry->name = name;
-    entry->adjust = tactyk_test__SET_REGISTER;
-    entry->test = tactyk_test__TEST_REGISTER;
-    entry->element_offset = ofs;
-    entry->array_offset = 0;
-}
-void tactyk_test__mk_xmm_register_test(char *name, uint64_t ofs) {
-    struct tactyk_test_entry *entry = tactyk_dblock__new_managedobject(base_tests, name);
-    entry->name = name;
-    entry->adjust = tactyk_test__SET_XMM_REGISTER_FLOAT;
-    entry->test = tactyk_test__TEST_XMM_REGISTER_FLOAT;
-    entry->element_offset = ofs;
-    entry->array_offset = 0;
-}
